requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.2.1
numpy==1.10.1
path.py==8.1.2
scipy==0.16.1
six==1.10.0
theanify==0.1.9
Theano==0.7.0
webargs==1.1.1
Werkzeug==0.11
wheel==0.24.0
